solved issue#2: added password validation using DRY approach
