+++ app/schemas/user_schemas.py
@@ -6,7 +6,7 @@ from enum import Enum
 from builtins import ValueError, any, bool, str
 from pydantic import BaseModel, EmailStr, Field, validator, root_validator
 from app.utils.nickname_gen import generate_nickname
-
+from app.utils.security import validate_password
 # User roles
 class UserRole(str, Enum):
     ANONYMOUS = "ANONYMOUS"
@@ -57,19 +57,11 @@ class UserBase(BaseModel):
 
 class UserCreate(UserBase):
     password: str = Field(..., example="Secure*1234")
-    @validator('password')
-    def validate_password(cls, v):
-        if len(v) < 8:
-            raise ValueError("Password must be at least 8 characters long.")
-        if not re.search(r"[A-Z]", v):
-            raise ValueError("Password must include at least one uppercase letter.")
-        if not re.search(r"[a-z]", v):
-            raise ValueError("Password must include at least one lowercase letter.")
-        if not re.search(r"[0-9]", v):
-            raise ValueError("Password must include at least one number.")
-        if not re.search(r"[!@#$%^&*(),.?\":{}|<>]", v):
-            raise ValueError("Password must include at least one special character.")
-        return v
+    @validator("password", pre=True, always=True)
+    def validate_password_field(cls, value):
+         if value:
+             validate_password(value)  # Raises a ValueError if invalid
+         return value
 
 
 class UserUpdate(UserBase):
